separate dns parser code into its own module
